requirements.txt
# Core dependencies
groq>=0.31.0
python-dotenv>=1.0.0

# API dependencies
fastapi>=0.116.1
uvicorn>=0.32.0

# Frontend dependencies
streamlit>=1.28.0

# Image processing
Pillow>=10.1.0

# Testing dependencies
requests>=2.31.0

python-multipart

# Frontend dependencies
streamlit
